requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
python-multipart>=0.0.9
google-genai>=1.0
openai>=1.0
anthropic>=0.25
ollama>=0.2
aiofiles>=23.0
